DMRC bot/requirements.txt
python-telegram-bot>=20,<22
APScheduler>=3.10
pytz
tzlocal
